+++ src/Components/Table/Table.js
@@ -27,7 +27,7 @@ const TableMain = styled.table`
   border-spacing: 0;
 `;
 const TableHead = styled.thead`
-  padding: 10px 0;
+  padding: 0px 0;
   border-bottom: 1px solid ${({ theme }) => theme.BORDER_PRIMARY};
   text-align: left;
 `;
@@ -41,14 +41,14 @@ const TableBody = styled.tbody``;
 const TableHeadData = styled.th`
   color: ${({ theme }) => theme.FONT_TITLE_INFO};
   font-weight: ${({ theme }) => theme.FONT_WEIGHT_NORMAL};
-  padding: 10px 20px;
+  ${({ alignHeaderText }) => alignHeaderText && `text-align: ${alignHeaderText}`};
+  padding: 10px 15px;
 `;
 const TableData = styled.td`
-  padding: 10px 20px;
-  text-align: left;
+  padding: 10px 15px;
+  text-align: ${({ center }) => center};
   display: ${({ copy }) => (copy ? 'flex' : '')};
   border: none;
-  min-width: 100px;
   text-decoration: ${({ skipped }) => skipped && 'line-through'};
   font-style: ${({ skipped }) => skipped && 'italic'};
 `;
@@ -127,8 +127,10 @@ const Table = ({
   );
 
   const buildTableHead = () =>
-    finalTableHeaders.map(({ displayName }) => (
-      <TableHeadData key={displayName}>{displayName}</TableHeadData>
+    finalTableHeaders.map(({ displayName, alignHeaderText = '' }) => (
+      <TableHeadData key={displayName} alignHeaderText={alignHeaderText}>
+        {displayName}
+      </TableHeadData>
     ));
 
   const buildSingleRow = (rowData, index) =>
@@ -173,6 +175,8 @@ const Table = ({
         hover = false,
         icon,
         skipped = false,
+        color = theme.FONT_LINK,
+        size = '1.4rem',
         copy = false,
       } = rowData[dataName] || {};
 
@@ -180,19 +184,21 @@ const Table = ({
       // Eg: value: [1456.43, 'vspn'] => {value[0]} {value[1]} (but use .map, since the array can vary in length)
       const finalValue = Array.isArray(value) ? value.map(singleValue => singleValue) : value;
       const valueMissing = value === '--' || value === '' || value === '--';
+      // If only displaying an icon, center it
+      const center = icon && value === '' ? 'center' : 'left';
 
       return (
-        <TableData title={hover || value} key={displayName} skipped={skipped} copy={copy}>
+        <TableData title={hover || value} key={displayName} skipped={skipped} copy={copy} center={center}>
           {link && !valueMissing && link !== pathname ? (
             <Link to={link}>
               {finalValue}
-              {icon && <Sprite icon={icon} size={'1.4rem'} color={theme.FONT_LINK} />}
+              {icon && <Sprite icon={icon} size={size} color={color} />}
               {copy && <CopyValue value={value} title={`Copy ${hover}`} />}
             </Link>
           ) : (
             <>
               {value}
-              {icon && <Sprite icon={icon} size={'1.4rem'} color={theme.FONT_LINK} />}
+              {icon && <Sprite icon={icon} size={size} color={color} />}
               {copy && <CopyValue value={value} title={`Copy ${hover}`} />}
             </>
           )}